[feature/personalbanker] - renamed the package names
