+++ app/src/main/AndroidManifest.xml
@@ -17,8 +17,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".branchManager.BranchManagerActivity"/>
-        <activity android:name=".personalBanker.PersonalBankerActivity"/>
+        <activity android:name=".branchmanager.BranchManagerActivity"/>
+        <activity android:name=".personalbanker.PersonalBankerActivity"/>
     </application>
 
 </manifest>
