+++ app/src/main/java/com/sam/samproject/viewmodels/LoginViewModel.java
@@ -11,9 +11,9 @@ import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.Toast;
 
-import com.sam.samproject.branchManager.BranchManagerActivity;
+import com.sam.samproject.branchmanager.BranchManagerActivity;
 
-import com.sam.samproject.personalBanker.PersonalBankerActivity;
+import com.sam.samproject.personalbanker.PersonalBankerActivity;
 
 public class LoginViewModel extends BaseViewModel {
     private String roles[];
